apps/api/pyproject.toml
[project]
name = "mello-api"
version = "0.0.1"
description = "Mello REST API — FastAPI on Cloud Run"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "firebase-admin>=6.6",
    "google-cloud-storage>=2.18",
    "python-dotenv>=1.0",
    "pydantic>=2.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "httpx>=0.27",
    "pytest-anyio>=0.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff.lint]
select = ["E", "F", "I"]
